Add memory pool and download parameters
Previous download implementation (with the growing size of download
chunk and download buffers the size of the number of chunks) used to
cause unbounded growth of memory usage and, therefore, restore process
failure.
New logic implies fixed chunk size and usage of memory pool the size
of the number of workers as download buffers. Chunk size and the number
of workers can be adjusted in config file.

- restore_max_concurrent_requests - int value to configure
  download concurrency (number of workers). Default: 6.
- restore_multipart_chunksize - bytesize format ("5B", "10KB", "1MB"
  etc)
  to specify download chunk size. Default: 10MB.